refactor/#67: Meeting Service CQRS패턴 반영
